requirements.txt
anthropic==0.34.2
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.24.0
